Switch tagging internals for compatibility with knowledgebase plugin, make contacts global, make projects belong to contacts.

+++ db/migrate/008_create_tags.rb
@@ -1,15 +1,22 @@
 class CreateTags < ActiveRecord::Migration
   def self.up
+    return if self.table_exists?("tags")
     create_table :tags do |t|
       t.column :name, :string
       t.timestamps
     end
     add_index :tags, :name
-
   end
   
 
   def self.down 
-    drop_table :tags
+    drop_table :tags if self.table_exists?("tags")
+  end
+  
+  #######
+  private
+  #######
+  def self.table_exists?(name)
+    ActiveRecord::Base.connection.tables.include?(name)
   end
 end
